move edge and node features to Trainer
